language: Replace selector rules with static selector parameters
In the first draft of the language there were many ways of deriving a dynamic value, including construct properties, selective construct definitions, function overloads, and selector rules - all of which could be used in combination!

To reduce this complexity, I've removed dynamic selector rules, and allowed function overloads to specify static selectors on their parameters. I think there is more work than can be done in this area, but it's a start!