game/tic-tac-toe: Make more use of functional features
In thinking about how the language will be used, I think it'll be the case that every program will have a very imperative 'core' that makes extensive use of varied and sometimes complex functional code. I modified this example to reflect the kind of thing I had in mind.

Note that what used to be `detect_winner()`, now `fn winner(Board board)`, does not modify the game state. This is now done in the main function instead.