compiler: Remove `Resolver::resolve_identity`
Though it is a helpful utility function, I have realised there are certain situations where we want to attempt to resolve an identity, but in some specific situations not emit an error if it isn't found.

Checking if the returned `optional` does or doesn't have a value is as much work as checking if the id is declared in scope, so I figured it best to factor it out.