compiler: Correctly parse whitespace and end of file
The way parsing was being done before meant the end of file would not always be handled properly. While fixing that, I discovered a bug in peek which meant it wasn't looking ahead of new lines as expected. While fixing these and it's knock on effects, I took the time to improve the parser slightly.