documentation: Split up 'overview.md' into reference guides
I realised that, at this point in the project, trying to maintain one large reference document with organised sections, concise writing, and solid explanations, is just a lot to ask - particularly given that all the details are in constant flux anyway. Splitting up individual features into their own files and providing more 'cut-and-dry' reference information I figure is probably better suited for iterating the language.

Eventually, I definitely want to create much more friendly documentation that would be helpful for someone looking to learn and/or get familiar with the language. That will probably be easier to do once the core of the language is better established though.