compiler: Replaced NamedPatterns with parameter Variables
I had implemented a notion of "Named Patterns", a way of declaring an identity associated with a pattern in the middle of a pattern expression. (Though, the implementation wasn't actually doing that).

I realised it was unclear exactly what that identity should represent, and that anything is could represent is already possible with either variables (in imperative contexts) or more clear pattern syntaxes (in the functional context).