compiler: Move declare method from APM to parser
I realised that different parts of the compiler will handle different errors in different ways, and that becomes difficult to manage if the APM throws any errors. Seen as declaration currently only happens in the parser anyway, I decided to move it.

As the compiler matures, I expect declaring intrinsic will no longer be declared by the parser, and so this aspect of it will be less strange.