compiler: Replace `Literal` with `IntrinsicValue`
I've been trying to figure out how patterns should work, and I realised there are times where it will be useful for the compiler to be able to internally reference _specific_ intrinsic values such as `3`, `true`, etc.

This is more specific than any literal that a user could write, so I replaced the `Literal` node with an `IntrinsicValue` one (which makes sense as this is how it is was being used anyway).