compiler: Remove 'expression patterns'
Previously, we would allow any kind of expression to be used as a pattern (e.g. on the lhs of a match rule). Parsing and resolution has now been rewritten so that there is a more intentional divide between the two. Namely, only intrinsic values can be used as patterns, and only in certain situations.