compiler: Separated resolver into resolver and checker
While working with some of the functions that determine and operate on expressions, it became clear to me that it would be helpful to have the resolver focus on resolving nodes with minimal pattern analysis. This will hopefully make it easier to maintain assumptions made in different parts of the program (for example `is_pattern_subset_of_superset` assumes that UnionPatterns never have exactly one child).

The resolver still needs to do _some_ pattern analysis so that it may generate helpful pattern hints (e.g. when resolving the patterns in a match statement). However, most of this work has now been moved into the checker.