compiler: Modify parsing functions
Previously, `eat` would return a token even if it was not the expected kind. This caused cascading compiler errors that were not easy to determine statically. So, I've rejigged `eat` to throw a complier error when the consumed token is not the correct type. This will mean a lot of the parsing functions have to handle errors properly now, so I've added the `check` function to help.