compiler: Changed method of creating spans in parser
I'm about to rework all of the parsing functions, and I realised the way spans are generated can be simplified (at least I think!). Given that the parsing functions are about to become a bit more involved now that invalid "eats" are compiler errors, I figured it made sense to make this simplification at the same time.