compiler: Basic call syntax parsing
Currently the resolver/checker gloss over the callee. I'm not sure how to represent it as an expression during parsing, but then resolve it as a procedure or entity type (i.e. constructible) for resolution and checking.

There are a few situations where any given piece of syntax could be a pattern, an dynamic expression, or potentially a static index (e.g. namespace.procedure_name). I think an important thing to tackle soon is how to handle those ambiguities?