compiler: Remove InvalidStatement node
`InvalidStatement` served no real purpose other than to have a `Statement` to return from `parse_statement` when no valid statement could be parsed (i.e. we never used it for providing useful error messages, etc). But, we were hardly even doing that, because we almost always peek for a statement before trying to parse it, meaning we almost never hit the code path where we create the `InvalidStatement`.