compiler: Inline statement parsing functions
This has some small benefits such as being able to reduce the amount of token peeking/confirming. However, my main hope is that this will make it easier to handle ambiguous syntaxes (e.g. if a statement begins with an expression, is it an expression statement, an assignment, an insert statement, a variable declaration, etc, etc - this is all easier to handle when all the required context exists in just one function)