compiler: Rework literal parsing
There is an ambiguity in the syntax which is that literals of various forms can be patterns or expressions depending on context. In some cases, we cannot know the context until we have already parsed the literal.

For example, the in case `Foo bar = 2`, when we peek the identity `Foo`, we don't know if we're looking at the pattern of a variable declaration, or the expression of an expression statement.

To resolve this, the parser will now generate a new node called `UnresolvedLiteral` whenever it is parsing something that could be an expression or could be a pattern. The resolver will then examine this, and determine which of the two it is.

When the resolver does so, it will output a `ExpressionLiteral` or `PatternLiteral` node, which contains the same span as the `UnresolvedLiteral` it replaces. The pattern or expression it contains then has no obligation to track it's own span.

This was quite a serious rework of the compiler so far. So, while I was at it, I also took the opportunity to reorganise `apm.h`, `apm.cpp`, and `apm-json.cpp`. The changes to these files are not actually as bad as they look. It was mostly a lot of rearranging code.

After this commit, the program now compiles again. However, I suspect there are many bugs! I am now going to track those down.