compiler: Rework variable declaration syntax
See game\card-attack\main.gambit for an example of the new sytnax.

The reworking of literal parsing was done with this change in mind. I'm not committed to this specific syntax, but I think I prefer it over `let`/`var`?

That said, not starting variable declarations with keywords AND not ending statements with semicolons can cause strange ambiguities! For example:

```
a = b
[Type] c
```

Was originally being parsed as

```
a = b[Type]
c
```

The next step is to refine the implementation of `parse_infix_expression_statement` to accurately parse the syntax and gracefully handle errors. (The current implementation sort of works, but is sloppy!)

After that, it would be a good idea to start addressing the numerous fixes and todos that have been building up. Now the the parser has had a bit of an overhaul, there is more work for the resolver and checker to be doing, which currently they are not.