compiler: Disable generator
My intention now is to have an intermediate representation between the checker and the generator, which takes the APM and converts it into an abstract representation of the C program that should be generated. The generator should essentially be a "reverse lexer", converting the IR into text, without having to worry about the differing semantics between languages.